requirements.txt
asgiref==3.8.1
attrs==24.1.0
Automat==22.10.0
certifi==2024.7.4
cffi==1.15.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==43.0.0
cssselect==1.2.0
defusedxml==0.7.1
Django==5.0.7
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
environ==1.0
filelock==3.15.4
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
itemadapter==0.9.0
itemloaders==1.3.1
jmespath==1.0.1
lxml==5.2.2
Markdown==3.6
mip==1.14.2
mysql==0.0.3
mysql-connector-python==9.0.0
mysqlclient==2.2.4
numpy==2.1.2
packaging==24.1
parsel==1.9.1
Protego==0.3.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyDispatcher==2.0.7
Pygments==2.18.0
PyJWT==2.9.0
pyOpenSSL==24.2.1
queuelib==1.7.0
requests==2.32.3
requests-file==2.1.0
Scrapy==2.11.2
service-identity==24.1.0
setuptools==72.1.0
six==1.16.0
snakeviz==2.2.0
sqlparse==0.5.1
tldextract==5.1.2
tornado==6.4.1
Twisted==24.3.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
w3lib==2.2.1
zope.interface==6.4.post2
